+++ docs/make.jl
@@ -17,6 +17,7 @@ makedocs(
     "Feature Matrix" => "feature_matrix.md",
     "Changelog" => "changelog.md",
     "Configuration" => "configuration.md",
+    "Programmatic API" => "programmatic_api.md",
     "Power-Flow Configuration" => "powerflow_configuration.md",
     "MATPOWER Import Configuration" => "matpower_import.md",
     "State-Estimation Configuration" => "state_estimation_configuration.md",